Fix issue with the way studentEnrollments.js was added
All the layouts for each entities blueprints weren't
loading because it was looking for the studentEnrollments.js
file in their relative static folder, but that file is
only in student_enrollments. Now the url_for is now pointing
to the studentEnrollments.js file from the static folder in
students_enrollments.